+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Changed
+- \#5 Inject odm manager via config and use alias @migo315 
+
 ## [0.0.2] - 2017-12-26
 ### Fixed
 - Fixed the uncaught mappingexception. @b3nl